feat: mving common to lib as rust standard approach
