--- src/constants.rs
@@ -1,14 +1,0 @@
-//! Constants
-pub const N_MEAN: f32 = 0.0;
-pub const N_STD_DEV: f32 = 1.0;
-pub const SQRT_2PI: f32 = 2.5066282;
-pub const HALF: f32 = 0.5;
-pub const DAYS_PER_YEAR: f32 = 365.25;
-pub use std::f32::consts::{E, PI};
-
-pub const A: f32 = 4.62627532e-01;
-pub const B: f32 = -1.16851917e-02;
-pub const C: f32 = 9.63541838e-04;
-pub const D: f32 = 7.53502261e-05;
-pub const _E: f32 = 1.42451646e-05;
-pub const F: f32 = -2.10237683e-05;
